move navbar buttons to its own module
