+++ client/src/components/Nav/index.js
@@ -1,21 +1,11 @@
-import React, { useContext } from 'react'
-import { Navbar, Button } from "react-bulma-components";
-import { AppContext } from '../../Context/AppProvider';
-import { tokenKey } from '../../hooks/config';
-const { Brand,
-	Item,
-	Container,
-} = Navbar;
+import React from 'react'
+import { Navbar } from "react-bulma-components";
+
+import { NavButtons } from './NavButtons';
+
+const { Brand, Item } = Navbar;
 
 export const Nav = () => {
-	const { tutorDetails, setTutorDetails, updateAppComponent } = useContext(AppContext)
-	const { loggedIn } = tutorDetails
-
-	const handleLogout = () => {
-		setTutorDetails({ loggedIn: false })
-		localStorage.removeItem(tokenKey)
-		updateAppComponent(null)
-	}
 
 	return (
 		<Navbar className="is-flex is-justify-content-space-between is-align-items-center">
@@ -28,34 +18,7 @@ export const Nav = () => {
 					/>
 				</Item>
 			</Brand>
-			<Container align="end">
-				{!loggedIn
-					? <>
-						< Button
-							size="small"
-							color="primary"
-							onClick={() => updateAppComponent('login')}
-						>
-							Login
-						</Button>
-						<Button
-							size="small"
-							color="primary"
-							onClick={() => updateAppComponent('signup')}
-						>
-							signup
-						</Button>
-					</>
-					: < Button
-						size="small"
-						color="primary"
-						onClick={handleLogout}
-					>
-						Logout
-					</Button>
-				}
-			</Container>
+			<NavButtons />
 		</Navbar >
 	);
 }
-
